+++ lib/recording_studio_api/api_definition.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "action_registry"
+require_relative "registered_endpoint_registry"
 require_relative "api_version_profile"
 require_relative "recordable_registry"
 
@@ -35,7 +36,7 @@ module RecordingStudioApi
                   :api_request_logging_enabled,
                   :api_request_logging_payload_mode,
                   :api_request_log_allowed_param_keys
-    attr_reader :name, :action_registry, :recordable_registry, :default_api_version, :api_version_profiles
+    attr_reader :name, :action_registry, :registered_endpoint_registry, :recordable_registry, :default_api_version, :api_version_profiles
 
     def initialize(name, defaults: nil)
       @name = name.to_s.freeze
@@ -53,6 +54,7 @@ module RecordingStudioApi
       @default_api_version = DEFAULT_API_VERSION
       @api_version_profiles = {}
       @action_registry = ActionRegistry.new
+      @registered_endpoint_registry = RegisteredEndpointRegistry.new
       @recordable_registry = RecordableRegistry.new
     end
 
@@ -88,6 +90,7 @@ module RecordingStudioApi
 
     def validate!
       action_registry.validate!
+      registered_endpoint_registry.validate!
       recordable_registry.validate!
       raise ConfigurationError, "authentication must be oauth for #{name}" unless authentication == :oauth
       raise ConfigurationError, "default_access must be read_only or read_write for #{name}" unless %i[read_only read_write].include?(default_access)
@@ -116,6 +119,7 @@ module RecordingStudioApi
         credential_ttl: credential_ttl,
         access_token_ttl: access_token_ttl,
         action_registrations: action_registry.to_h,
+        registered_endpoint_registrations: registered_endpoint_registry.to_h,
         recordable_registrations: recordable_registry.to_h
       }
     end